requirements.txt
fastapi
uvicorn
gunicorn
redis
pydantic
piper-tts